Fix camera/flash synchronization
Higher FPS used for the camera, dropping the frames taken before the LED
was updated.
I should probably clean up some of the atomic state updates, but it
works for now.
Note that this code works because System.nanoTime() *happens* to return
the ns since the system turned on, which it is not guaranteed to do.