Work on tiled diffs
Considering that a Java boolean is 1 byte (=8 bits), storing 64 bools in
a long should reduce memory const by up to 90%.
This change might not even come at the cost of CPU, because we can do
some bit operations on longs (we do a bunch of row/col testing) that we
can't do on booleans.
This class needs a better name, though.